refactor(#57): 유즈케이스 모듈 분리
