--- app/src/main/java/com/ssafy/neegongnaegong/domain/model/GitHubRepo.kt
@@ -1,8 +1,0 @@
-package com.ssafy.neegongnaegong.domain.model
-
-data class GitHubRepo(
-    val id: Int,
-    val name: String,
-    val description: String,
-    val url: String?
-)
