--- app/src/main/java/com/ssafy/neegongnaegong/domain/repository/GitHubRepository.kt
@@ -1,8 +1,0 @@
-package com.ssafy.neegongnaegong.domain.repository
-
-import com.ssafy.neegongnaegong.domain.model.GitHubRepo
-
-interface GitHubRepository {
-
-    suspend fun getUserRepos(username: String): List<GitHubRepo>
-}
