--- app/src/main/java/com/ssafy/neegongnaegong/domain/usecase/GetUserReposUseCase.kt
@@ -1,11 +1,0 @@
-package com.ssafy.neegongnaegong.domain.usecase
-
-import com.ssafy.neegongnaegong.domain.model.GitHubRepo
-import com.ssafy.neegongnaegong.domain.repository.GitHubRepository
-
-class GetUserReposUseCase(private val repository: GitHubRepository) {
-
-    suspend operator fun invoke(username: String): List<GitHubRepo> {
-        return repository.getUserRepos(username)
-    }
-}
