+++ app/src/main/java/com/ssafy/neegongnaegong/module/di/RepositoryModule.kt
@@ -2,12 +2,10 @@ package com.ssafy.neegongnaegong.module.di
 
 import com.ssafy.neegongnaegong.data.repository.AuthRepositoryImpl
 import com.ssafy.neegongnaegong.data.repository.CalendarRepositoryImpl
-import com.ssafy.neegongnaegong.data.repository.GitHubRepositoryImpl
 import com.ssafy.neegongnaegong.data.repository.StudiesRepositoryImpl
 import com.ssafy.neegongnaegong.data.repository.UserRepositoryImpl
 import com.ssafy.neegongnaegong.domain.repository.AuthRepository
 import com.ssafy.neegongnaegong.domain.repository.CalendarRepository
-import com.ssafy.neegongnaegong.domain.repository.GitHubRepository
 import com.ssafy.neegongnaegong.domain.repository.StudiesRepository
 import com.ssafy.neegongnaegong.domain.repository.UserRepository
 import dagger.Binds
@@ -19,34 +17,19 @@ import javax.inject.Singleton
 @Module
 @InstallIn(SingletonComponent::class)
 interface RepositoryModule {
+    @Binds
+    @Singleton
+    fun bindAuthRepository(authRepositoryImpl: AuthRepositoryImpl): AuthRepository
 
     @Binds
     @Singleton
-    fun bindGitHubRepository(
-        gitHubRepositoryImpl: GitHubRepositoryImpl
-    ): GitHubRepository
+    fun bindUserRepository(userRepositoryImpl: UserRepositoryImpl): UserRepository
 
     @Binds
     @Singleton
-    fun bindAuthRepository(
-        authRepositoryImpl: AuthRepositoryImpl
-    ): AuthRepository
+    fun bindStudiesRepository(studiesRepositoryImpl: StudiesRepositoryImpl): StudiesRepository
 
     @Binds
     @Singleton
-    fun bindUserRepository(
-        userRepositoryImpl: UserRepositoryImpl
-    ): UserRepository
-
-    @Binds
-    @Singleton
-    fun bindStudiesRepository(
-        studiesRepositoryImpl: StudiesRepositoryImpl
-    ): StudiesRepository
-
-    @Binds
-    @Singleton
-    fun bindCalendarRepository(
-        calendarRepositoryImpl: CalendarRepositoryImpl
-    ): CalendarRepository
+    fun bindCalendarRepository(calendarRepositoryImpl: CalendarRepositoryImpl): CalendarRepository
 }
